Tasks 1 to 6 updated Tasks 7 and 9 added

+++ src/Task2/Ex1/Calculator.java
@@ -23,7 +23,7 @@ public class Calculator {
     }
 
     public long Subtract(int a, int b) {
-        return ((long)(a-b));
+        return ((a-b));
     }
 
     public double Multiply(double a, double b) {
